Fix: fix null reaction, add seed script
Due to the use of the push array method, a null piece of data was
making it's way into the reactions to be pushed to the db. simply made
the reactions = to the generated data.

Fixed incorrect error and success messages in controllers.

Added in the backup localhost storage for the db.

Changed the name of the seeds folder to utils and created a short
npm script to seed the db.